feat: fix tests, update connexion for strict checking
Loading the application failed as connexion expects the operationId to
be a fully resolvable function path. These have all been updated and
the test load succeeds.

Connexion setup no longer attempts to load the swagger_ui as that wasn't
compatible with the entire OpenAPI 2.0 spec.

Strict request validation is now enabled, and the request will be
rejected if the validation fails with the appropriate errors.