feat: refactor helper functions, add more thorough error handling
Refactors the helper functions to a customer.py module for individual
testing as well as removing any dependency on Flask state.

Added type hints to all customer helper functions and most of the API
methods.

Added error handling for Stripe errors to Flask itself for catching.
This removes duplicate handling inside functions and provides a uniform
place to translate errors as appropriate.