bugfix Customer Subscription Webhooks - Deleted Customer
We were encountering 500 errors on some Customer Subscription Deleted events. When investigating, the fetch Customer calls were returning 200s, but there was a distinct possibility that the Customer record being returned was for a deleted entity, which would result in a failed attempt to create the payload. I added checks to verify if a Customer has been deleted to the get_user_id methods and added addtional logging information. These deleted Customer checks have been added to all of the subscription events.